# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "ToDo API",
      "version": "1.0.0"
    },
    "paths": {
      "/todos": {
        "get": {
          "summary": "Get all Todos",
          "responses": {
            "200": {
              "description": "A list of todos",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Todo"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new Todo",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Todo created"
            }
          }
        }
      },
      "/todos/{id}": {
        "get": {
          "summary": "Get a specific Todo",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "Todo ID",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Todo details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        },
        "put": {
          "summary": "Update an existing Todo",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "Todo ID",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Todo updated"
            }
          }
        },
        "delete": {
          "summary": "Delete a Todo",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "Todo ID",
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Todo deleted"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Todo": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "completed": {
              "type": "boolean"
            }
          }
        }
      }
    }
  }
  